refactor: Cut down main's size
